renaming generate_images_from_random_prompt to generate_images_from_prompt_generator
